Conflicts:
	entity-management-common/src/main/java/eu/europeana/entitymanagement/config/EMSettings.java
	entity-management-common/src/main/java/eu/europeana/entitymanagement/config/EntityManagementConfigurationImpl.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/ConceptScheme.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/impl/BaseAggregation.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/impl/BaseConcept.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/impl/BaseConceptScheme.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/impl/BaseEntity.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/impl/BaseEntityRecord.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/impl/BaseTimespan.java
	entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/mongo/impl/EntityRecordImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/EMController.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/RdfDepiction.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/RdfResource.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlAddressImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlAgentImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlAggregationImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlBaseEntityImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlConceptImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlConstants.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlIsAggregatedByImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlMultilingualString.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlOrganizationImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlPlaceImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlTimespanImpl.java
	entity-management-web/src/main/java/eu/europeana/entitymanagement/web/xml/model/XmlWebResourceImpl.java
	entity-management-web/src/main/resources/entitymanagement.properties
	entity-management-web/src/test/java/eu/europeana/entitymanagement/mongo/InsertAndRetrieveEntityToAndFromMongoDB.java

+++ entity-management-definitions/src/main/java/eu/europeana/entitymanagement/definitions/model/Aggregation.java
@@ -20,9 +20,9 @@ public interface Aggregation {
 
     int getRecordCount();
 
-    void setPageRank(int pageRank);
+    void setPageRank(double pageRank);
 
-    int getPageRank();
+    double getPageRank();
 
     void setModified(Date modified);
